content-factory/requirements.txt
google-genai
alembic
sqlalchemy
psycopg2-binary
beautifulsoup4
httpx
pillow
